project-root/chatbot/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
mistralai==0.0.12
python-dotenv==1.0.0
langchain==0.0.350
langchain-mistralai==0.0.1
langchain-community==0.0.10
tiktoken==0.5.1
requests==2.31.0 